Move Plan types to planner/plan.ts
This allows to centralize how a Node is created and the id is generated

Change-type: minor
